BookSwap/requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
django-softdelete==0.11.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1